+++ src/JK.Common/Extensions/DateTimeOffsetExtensions.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using JK.Common.TypeHelpers;
 
 namespace JK.Common.Extensions;
 
@@ -7,45 +6,40 @@ namespace JK.Common.Extensions;
 public static class DateTimeOffsetExtensions
 {
     /// <summary>Adds given number of business days to a date.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
+    /// <param name="date">Current DateTimeOffset object from extension method.</param>
     /// <param name="days">Number of days to add (can be negative).</param>
     /// <returns>The date the given amount of business days from the start date.</returns>
-    public static DateTimeOffset AddWorkDays(this DateTimeOffset dt, in int days) => DateHelper.AddWorkDays(dt, days);
+    public static DateTimeOffset AddWorkDays(this DateTimeOffset date, in int days)
+    {
+        var direction = days < 0 ? -1 : 1;
+        var daysToAdd = Math.Abs(days);
+        var count = 0;
+        while (count < daysToAdd)
+        {
+            date = date.AddDays(direction);
+            if (IsWeekday(date))
+            {
+                count++;
+            }
+        }
 
-    /// <summary>Gets the abbreviated day name.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
-    /// <returns>Abbreviated day name.</returns>
-    public static string GetAbbreviatedDayName(this DateTimeOffset dt) => DateHelper.GetAbbreviatedDayName(dt);
-
-    /// <summary>Gets the abbreviated month name.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
-    /// <returns>Abbreviated month name.</returns>
-    public static string GetAbbreviatedMonthName(this DateTimeOffset dt) => DateHelper.GetAbbreviatedMonthName(dt);
-
-    /// <summary>Gets the full name of a given day.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
-    /// <returns>Full day name.</returns>
-    public static string GetDayName(this DateTimeOffset dt) => DateHelper.GetDayName(dt);
-
-    /// <summary>Gets the full name of a month.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
-    /// <returns>Full month name.</returns>
-    public static string GetMonthName(this DateTimeOffset dt) => DateHelper.GetMonthName(dt);
+        return date;
+    }
 
     /// <summary>Determines whether or not a given date is between (inclusive) the given start and end dates.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
+    /// <param name="date">Date to check</param>
     /// <param name="start">Start of date range to check</param>
     /// <param name="end">End of date range to check</param>
     /// <returns>True if date falls within range, otherwise false</returns>
-    public static bool IsBetween(this DateTimeOffset dt, in DateTimeOffset start, in DateTimeOffset end) => DateHelper.IsBetween(dt, start, end);
+    public static bool IsBetween(this DateTimeOffset date, in DateTimeOffset start, in DateTimeOffset end) => start <= date && end >= date;
 
     /// <summary>Determines if given date is a weekday.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
+    /// <param name="date">Current DateTimeOffset object from extension method.</param>
     /// <returns>True if is a weekday, otherwise false.</returns>
-    public static bool IsWeekday(this DateTimeOffset dt) => DateHelper.IsWeekday(dt);
+    public static bool IsWeekday(this DateTimeOffset date) => date.DayOfWeek.IsWeekday();
 
     /// <summary>Determines if given date is a weekday.</summary>
-    /// <param name="dt">Current DateTimeOffset object from extension method.</param>
+    /// <param name="date">Current DateTimeOffset object from extension method.</param>
     /// <returns>True if is a weekend, otherwise false.</returns>
-    public static bool IsWeekend(this DateTimeOffset dt) => DateHelper.IsWeekend(dt);
+    public static bool IsWeekend(this DateTimeOffset date) => !date.DayOfWeek.IsWeekday();
 }